Verify PR temp comments
