+++ alibi_detect/utils/saving/__init__.py
@@ -2,9 +2,11 @@
 # ###### these comments should be removed prior to PR merge. #########
 # ####################################################################
 #
-# 1. # Moved `saving` from `/utils` to here: `/utils/saving/`. `save_detector` and `load_detector` are
-# `import_optional`'ed in `/utils/saving/__init__.py`.
-# TODO: config-driven-detectors moves this module out of utils. So will require refactoring!
+# 2. Moved `saving` from `/utils/saving.py` to here: `/utils/saving/`.
+# `save_detector` and `load_detector` are `import_optional`'ed in
+# `/utils/saving/__init__.py`.
+# TODO: config-driven-detectors moves this module out of utils. So
+#  will require refactoring!
 
 
 from alibi_detect.utils.missing_optional_dependency import import_optional
